left, right, extendLeft, extendRight
a more symmetric API. previously it was only possible to extend towards
the right, now we can extend in both directions. furthermore, we can now
choose whether we want to fail if there are no elements in that
direction (using 'left' and 'right'), or if we want to succeed no matter
what, at the cost of making the update function a bit more complicated
because it now has to deal with the case in which there are no more
elements.