runOracleMachineWithAnswers
The old 'runOracleMachineWithReplacements' always succeeded because it
used the questions as answers if it ran out of replacements. The
downside is that the questions and the answers must have the same type.

The new 'runOracleMachineWithAnswers' does allow questions and answers
to have different types, but the transformation fails if we run out of
answers.