replace OracleMachine with SplitTraversable
I was planning to represent a @Dissected t c a j@ using a
'DissectedList' of 'Suspended' computations:

    newtype Dissected t c a j = Dissected
      { unDissected :: DissectedList (Suspended c c j (t c))
                                     (Suspended c a j (t c))
                                     j
      }

This representation is type-safe:
- I can access the @c@s as @DissectedList.clowns . each . Suspended.focus@
- I can access the @a@ as @DissectedList.focus . Suspended.focus@
- I can access the @j@s as @DissectedList.jokers@
- If @a@ is @j@, I can move left by adding the focus to the beginning of
  the jokers and moving the last @Suspended c c j (t c)@ into focus
- If @a@ is @c@, I can move right by adding the focus to the end of the
  @Suspended c c j (t c)@s and using that @c@ to resume that same
  Suspended computation in order to obtain a new Suspended computation.
- If we reach the end of the 'OracleMachine' computation, we obtain a
  @t c@.

Unfortunately, this representation has a fatal flaw: it bakes in a
particular type for the clowns and the jokers, thereby making it
impossible to implement operations which change the type of 'c' and 'j'.
In particular, I want to represent 'NonEmpty' as

    newtype NonEmpty t a = NonEmpty
      { unNonEmpty :: Dissected t Void a Void }

Which would be easy if we could traverse the clowns and the jokers with
@\_ -> empty@, but is instead impossible.

So I am switching to a slightly less safe representation,
'SplitTraversable', in which we store the shape of the 'Traversable'
separately from its values. This relies on the invariant that traversing
the same 'Traversable' more than once always traverses the same number
of elements. This is a free theorem, so we can rely on it, but the types
don't guarantee it, so we'll have to throw in a call to 'fromJust' at
some point. Oh well.